Fix gcs
